requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-SocketIO==5.1.1
gevent
gevent-websocket==0.10.1
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.23
WTForms==2.3.3
email-validator==1.1.3